refactor: to put the resolver logic in a separate layer
