+++ src/routes/graphql/SchemaEntities/Mutations/Mutation.ts
@@ -1,4 +1,3 @@
-import fetch from 'node-fetch';
 import { GraphQLObjectType } from 'graphql';
 import {
 	addPostDtoInput,
@@ -17,11 +16,7 @@ import {
 	ProfileType,
 	UserType,
 } from '../Querys/QueryDefs';
-import { UserEntity } from '../../../../utils/DB/entities/DBUsers';
-import { ProfileEntity } from '../../../../utils/DB/entities/DBProfiles';
-import { PostEntity } from '../../../../utils/DB/entities/DBPosts';
-import { MemberTypeEntity } from '../../../../utils/DB/entities/DBMemberTypes';
-import { routeUrl } from '../../config';
+import { userService } from './../../services/UserService';
 
 export const Mutation = new GraphQLObjectType({
 	name: 'Mutation',
@@ -29,197 +24,47 @@ export const Mutation = new GraphQLObjectType({
 		addUser: {
 			type: UserType,
 			args: { input: { type: addUserDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Omit<UserEntity, 'id'>>
-			) {
-				const { firstName, lastName, email } = input;
-				const body = JSON.stringify({ firstName, lastName, email });
-				const response = await fetch(routeUrl.users, {
-					method: 'POST',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.addUser,
 		},
 		addProfile: {
 			type: ProfileType,
 			args: { input: { type: addProfileDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Omit<ProfileEntity, 'id'>>
-			) {
-				const {
-					avatar,
-					sex,
-					birthday,
-					country,
-					street,
-					city,
-					userId,
-					memberTypeId,
-				} = input;
-				const body = JSON.stringify({
-					avatar,
-					sex,
-					birthday,
-					country,
-					street,
-					city,
-					userId,
-					memberTypeId,
-				});
-				const response = await fetch(routeUrl.profiles, {
-					method: 'POST',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.addProfile,
 		},
 		addPost: {
 			type: PostType,
 			args: { input: { type: addPostDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Omit<PostEntity, 'id'>>
-			) {
-				const { content, title, userId } = input;
-				const body = JSON.stringify({ content, title, userId });
-				const response = await fetch(routeUrl.posts, {
-					method: 'POST',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.addPost,
 		},
 		updateUser: {
 			type: UserType,
 			args: { input: { type: updateUserDtoInput } },
-			async resolve(parent: unknown, { input }: Record<'input', UserEntity>) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(`${routeUrl.users}/${input.id}`, {
-					method: 'PATCH',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.updateUser,
 		},
 		updateProfile: {
 			type: ProfileType,
 			args: { input: { type: updateProfileDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Omit<ProfileEntity, 'userId'>>
-			) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(`${routeUrl.profiles}/${input.id}`, {
-					method: 'PATCH',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.updateProfile,
 		},
 		updatePost: {
 			type: PostType,
 			args: { input: { type: updatePostDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Omit<PostEntity, 'userId'>>
-			) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(`${routeUrl.posts}/${input.id}`, {
-					method: 'PATCH',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.updatePost,
 		},
 		updateMemberType: {
 			type: MemberType,
 			args: { input: { type: updateMemberDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', MemberTypeEntity>
-			) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(`${routeUrl.memberTypes}/${input.id}`, {
-					method: 'PATCH',
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					body,
-				});
-
-				return await response.json();
-			},
+			resolve: userService.updateMemberType,
 		},
 		subscribeUser: {
 			type: UserType,
 			args: { input: { type: subscribeUserDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Pick<ProfileEntity, 'id' | 'userId'>>
-			) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(
-					`${routeUrl.users}/${input.id}/subscribeTo`,
-					{
-						method: 'POST',
-						headers: {
-							'Content-Type': 'application/json',
-						},
-						body,
-					}
-				);
-
-				return await response.json();
-			},
+			resolve: userService.subscribeUser,
 		},
 		unSubscribeUser: {
 			type: UserType,
 			args: { input: { type: unSubscribeUserDtoInput } },
-			async resolve(
-				parent: unknown,
-				{ input }: Record<'input', Pick<ProfileEntity, 'id' | 'userId'>>
-			) {
-				const body = JSON.stringify({ ...input });
-				const response = await fetch(
-					`${routeUrl.users}/${input.id}/unsubscribeFrom`,
-					{
-						method: 'POST',
-						headers: {
-							'Content-Type': 'application/json',
-						},
-						body,
-					}
-				);
-
-				return await response.json();
-			},
+			resolve: userService.unSubscribeUser,
 		},
 	},
 });
